fix: change publicShares column type from int to bigint for NPC corporation share counts
Agent-Logs-Url: https://github.com/garveen/eve-sde-converter/sessions/332d74a5-6b8f-4bfc-ac3e-8798de62f90c

+++ src/schema.sql
@@ -420,7 +420,7 @@ CREATE TABLE `crpNPCCorporations` (
   `investorShares4` int(11) DEFAULT NULL,
   `friendID` int(11) DEFAULT NULL,
   `enemyID` int(11) DEFAULT NULL,
-  `publicShares` int(11) DEFAULT NULL,
+  `publicShares` bigint DEFAULT NULL,
   `initialPrice` int(11) DEFAULT NULL,
   `minSecurity` float DEFAULT NULL,
   `scattered` tinyint(1) DEFAULT NULL,
